Possible Fix: SIGPIPE TestRunningViewController fatal Signal 13, Code 0

+++ ooniprobe/Utility/SettingsUtility.h
@@ -31,7 +31,7 @@
 
 + (void)incrementAppOpenCount;
 
-+ (long)getAppOpenCount;
++ (NSInteger)getAppOpenCount;
 
 + (void)registeredForNotifications;
 @end
